rename: global -> globalCss
